@@ -8,7 +8,7 @@
   },
   "devDependencies": {
     "@sveltejs/kit": "next",
-    "svelte": "^4.0.0",
+    "svelte": "^5.0.0",
     "@sveltejs/adapter-vercel": "next"
   },
   "type": "module",
