@@ -1,9 +1,11 @@
 <script>
 	import '../app.css';
+	/** @type {{children?: import('svelte').Snippet}} */
+	let { children } = $props();
 </script>
 
 <main>
-	<slot />
+	{@render children?.()}
 </main>
 
 
